requirements.txt
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.41.0
pyyaml>=6.0.1
numpy>=1.24.0
pillow>=10.0.0
scikit-learn>=1.3.0
tqdm>=4.66.0
tensorboard>=2.14.0
